feat: implement custom hook for owl tips and clean up related code
